build: move breaking changes script away from gulp
The script that logs out the breaking changes that need to be cleaned up was barely using Gulp anyway. These changes make it so it doesn't depend on Gulp and the `package-tools`.
